requirements.txt
google-genai
typer
rich
python-dotenv